Deleting the list items selected
