+++ src/modules/remove.js
@@ -1,13 +1,22 @@
-const listContainer = document.querySelector('.listContainer');
+let todoTasks;
+
 const remove = (removeitem, array) => {
-    console.log(typeof parseInt(removeitem.id-1))
-    array.splice(parseInt(removeitem.id-1), 1);
+    if (localStorage.getItem('todoList') === null) {
+        todoTasks = [];
+      } else {
+        todoTasks = JSON.parse(localStorage.getItem('todoList'));
+      }
+
     for(let i=parseInt(removeitem.id-1); i < array.length; i++){
         array[i].index--;
     }
+    
     removeitem.remove();
-
+    array.splice(parseInt(removeitem.id-1), 1);
+    localStorage.setItem('todoList', JSON.stringify(array));
+    
     return array;
+    
 }
 
 export default remove;
